Replace home page video loop
Closes #51
