--- components/HiveVideo/index.ts
@@ -1,1 +1,0 @@
-export { HiveVideo } from './HiveVideo'
