+++ components/HomeLoop/index.ts
@@ -1,0 +1,1 @@
+export { HomeLoop } from './HomeLoop'
